+++ Safuan_Ysupov/Calculater.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30503.244
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Calculater", "Calculater.vcxproj", "{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Debug|x64.ActiveCfg = Debug|x64
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Debug|x64.Build.0 = Debug|x64
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Debug|x86.ActiveCfg = Debug|Win32
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Debug|x86.Build.0 = Debug|Win32
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Release|x64.ActiveCfg = Release|x64
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Release|x64.Build.0 = Release|x64
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Release|x86.ActiveCfg = Release|Win32
+		{E1126AAE-7D84-4AB8-91E8-B26B84D0F22B}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {349B4C95-3607-4DF1-9A33-3C4447B25B5C}
+	EndGlobalSection
+EndGlobal
