+++ Safuan_Ysupov/calc_cin.txt
@@ -1,0 +1,11 @@
+4+1; 
+let x=4;
+let y=3;
+(x+y)*20; 
+x%10;
+4+11-4*(2+12);
+let n=65;
+y/x;
+43+23-n*5-2/(x+y)/23*1002*(12-4*(1.4-5.7+x-y*4));
+5%3;
+q
