+++ Safuan_Ysupov/calc_cout.txt
@@ -1,0 +1,11 @@
+> = 5
+> = 4
+> = 3
+> = 140
+> = 4
+> = -41
+> = 65
+> = 0.75
+> = -1020.77
+> = 2
+>
